pyproject.toml
[project]
name = "fantasy-football-analysis"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "nflreadpy>=0.1.5",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyarrow>=23.0.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.40.0",
]
